test: add refersto, generate schema tests
